Issue: WSW-17509 MISRA defects in Component midware_wifi.
17021251 (Mandatory issue) :- MISRA C-2012 rule 21.18,The size_t argument passed
to any function shall have an appropriate value.

The length argument with possible value 0, which is passed to function "memcpy", shall have a positive
value that is no greater than the size of the smallest object passed to the function through a pointer parameter.

Change Description :- Check the length argument before calling the memcpy function, it's always not smaller than the zero and
not greter then the source length.

Test:- Compile and Build successful with RT1060 & 8978 wifi_cli.
Successfully scan,connect with the Ex-AP and ping.

Reviewers: Abhinav K,Sandip Dalvi, Supriya Inamdar,Abhishek Misra.

Signed-off-by: Ayush Nema <ayush.nema@nxp.com>