Issue: WSW-17509 MISRA Rule 8.6 patch 5
MISRA C-2012 Compliance for Rule 8.6 Function is declare but never define.

Change Description :- All the declaration already having the definition but
covered by the #if, now declaration is also covered by the same and if declaration is only there then deleted this.

Test:- Compile and Build successful with RT1060 & 8978 wifi_cli.
Successfully scan,connect with the Ex-AP and ping.

Reviewers: Abhinav K,Sandip Dalvi, Supriya Inamdar,Abhishek Misra.

Signed-off-by: Ayush Nema <ayush.nema@nxp.com>