Issue: WSW-17509 MISRA C-2012 Rule 10.4 patch 20
Defects in Component midware_wifi.
After analyzing using Coverity and Misra tools, MISRA C-2012 rule 10.4,
essential type of the left hand operand is not the same as that of the right operand.

Change Description: Changes done as to make left hand operand is same as that of the
right operand. mlan_act_ioctl enum related changes.

Test: Compile and Build successful with RT1060 & 8987 wifi_cli.
Checked for basic functionality.

Reviewers: Abhishek Misra, Abhinav K, Supriya Inamdar, Sandip Dalvi.

Signed-off-by: Sandeep Mane <sandeep.mane@nxp.com>