Issue: WSW-17509 MISRA C-2012 Rule 11.2
[Issue]
Violation of MISRA C-2012 Rule 11.2: Conversions shall not be performed between
a pointer to an incomplete type and any other type.

xTaskGetCurrentTaskHandle() returns "xTaskHandle_t *" and it is assigned to
"handle" variable which is of type "xTaskHandle_t **".

[Fix]
Changed type of "handle" var to single ptr instead of double ptr as double ptr
is not required.

[Test]
1. Build success of wifi_cli app for RT1060 on IAR, Keil and MCUX.
2. Checked if no new warnings/issues are reported by Coverity.
3. Sanity test of wifi_cli app in STA and uAP mode.

[Open Issues/Todo]
NA

[Reviewers]
Abhishek Misra, Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>