Issue: WSW-19604 [Embedded]uAP beacons intermittently when started with ACS in BGN40
[Issue]
DUT as AP does not send beacons periodically as expected, if channel number is not provided.

[Fix]
Making the channel number a mandatory parameter to setup uAP using wlan-add command.

[Validation done]
If channel number is not provided as argument, the wlan-add command fails.
If channel number is provided as 0, the wlan-add command fails.
Setting up uAP is successful with a non-zero channel number(36) defined during AP configuration.
Setting up DUT as STA works normally as before and is not affected by changes.

[Open Issues/Todo]
Changes made for this commit are temporary. To be removed once ACS related fix is added in firmware.

[Reviewers]
Abhinav Kulkarni, Sandip Dalvi, Supriya Inamdar

Signed-off-by: Keval Thakker <keval.thakker@nxp.com>