Issue: WSW-17509 MISRA C-2012 Rule 14.3
[Issue]
Violation of MISRA C-2012 Rule 14.3: Controlling expressions shall not be invariant

If a controlling expression has an invariant value, it is possible that there is a programming error.
Any code that cannot be reached due to the presence of an invariant expression may be removed by the
compiler. This might have the effect of removing defensive code, for instance, from the executable.

[Fix]
Removed unreachable code

[Test]
1. Build success of wifi_cli app for RT1060 on IAR, Keil and MCUX.
2. Checked if no new warnings/issues are reported by Coverity.
3. Sanity test of wifi_cli app in STA and uAP mode.

[Open Issues/Todo]
NA

[Reviewers]
Abhishek Misra, Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>