Issue: WSW-17509 MISRA C-2012 Rule 10.1 Patch 1
[Issue]
Violation of MISRA C-2012 Rule 10.1: Operands shall not be of an inappropriate
essential type.

An operand of essentially enum type should not be used in an arithmetic
operation because an enum object uses an implementation-defined integer type.
An operation involving an enum object may therefore yield a result with an
unexpected type.

[Fix]
Changed wmerror enum to macros. It fixes 200+ issues

[Test]
1. Build success of wifi_cli app for RT1060 on IAR, Keil and MCUX.
2. Checked if no new warnings/issues are reported by Coverity.
3. Sanity test of wifi_cli app in STA and uAP mode.

[Open Issues/Todo]
NA

[Reviewers]
Abhishek Misra, Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>