Issue: WSW-17509 MISRA C-2012 Rule 10.1 Patch 3
[Issue]
Violation of MISRA C-2012 Rule 10.1: Operands shall not be of an inappropriate
essential type.

An operand of essentially enum type should not be used in an arithmetic
operation because an enum object uses an implementation-defined integer type.
An operation involving an enum object may therefore yield a result with an
unexpected type.

[Fix]
if condition updated to be compliant with rule

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.

[Open Issues/Todo]
Other 10.1 issues

[Reviewers]
Sanjay Yadav, Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni

Signed-off-by: Abhishek Misra <abhishek.misra@nxp.com>