Issue: WSW-17509 MISRA Rule 15.6 and 15.7 patch
[Issue]
Rule 15.6 The body of an iteration-statement or a selection-statement shall be a compound statement..
Rule 15.7 All if ... else if constructs shall be terminated with an else statement.

[Fix]
Format if-else as per rule.

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.

[Open Issues/Todo]

Below issues are False Positive. They are not see on code inspection.
9474641		dhcpd\dhcp-server.c
9475141		dhcpd\dhcp-server.c
9476871		dhcpd\dhcp-server.c
9476938		dhcpd\dhcp-server.c
9477798		dhcpd\dhcp-server.c
9478346		dhcpd\dhcp-server.c
10045207    	dhcpd\dhcp-server.c
11768249	dhcpd\dhcp-server.c

Below issues are coming from LWIP Module
9474735		port\lwip\wifi_netif.c
9477447		port\lwip\wifi_netif.c
9478437		port\lwip\wifi_netif.c

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

Signed-off-by: Abhishek Misra <abhishek.misra@nxp.com>