Issue: WSW-19613 [RT1170][9177][MMH]No provision to configure uAP in VHT-20MHz and VHT-40MHz
[Issue]
New reqirement to support configuing uAP in VHT-20MHz and VHT-40MHz.

[Fix]
Added addtional option 3 in wlan-set-uap-bandwidth to set bandwidth to 80MHz.
Added logic so that VHT BW follows VHT-caps if 80MHz is set by user else it follows HT-caps for 20 and 40 MHz cases.
ACS and setting uAP channel to 14 or less will not be allowed in case user set BW to 80MHz.

[Validation done]
Setting up uAP is successful for BW set to 40/20 MHz and channel is set automatically or to any other valid value.
Setting up uAP is successful for BW set to 80MHz and channel is set to 36/44/48.
Setting up uAP is not allowed if BW set to 80MHz and channel is set to 14 or less or selected automatically.

[Open Issues/Todo]
None.

[Reviewers]
Sandip Dalvi, Abhinav Kulkarni, Supriya Inamdar

Signed-off-by: Keval Thakker <keval.thakker@nxp.com>