Issue: WSW-17509 MISRA wifi_nxp update Patch 4.0 fn
[Issue]
New MISRA issues have been introduced in function wifi_handle_fw_event().

[Fix]
Fixes for below rules.

Rule 10.3 The value of an expression shall not be assigned to an object with a narrower essential type or of a different essential type category.

Rule 10.4 Both operands of an operator in which the usual arithmetic conversions are performed shall have the same essential type category.

Rule 10.7 If a composite expression is used as one operand of an operator in which the usual arithmetic conversions are performed then the other operand shall not have wider essential type.

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

Signed-off-by: Abhishek Misra <abhishek.misra@nxp.com>