Issue: WSW-17509 MISRA Fixes for WLCMGR patch 1
[Issue]

Rule 10.1 Operands shall not be of an inappropriate essential type.
Rule 10.4 Both operands of an operator in which the usual arithmetic conversions are performed shall have the same essential type category.
Rule 22.8 The value of errno shall be set to zero prior to a call to an errno-setting-function.
Rule 22.9 The value of errno shall be tested against zero after calling an errno-setting-function.

Also remove CLI and WLCMGR dependency on MFALSE/MTRUE

[Fix]
Comply as per rules

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

Signed-off-by: Abhishek Misra <abhishek.misra@nxp.com>