Issue: WSW-17509 MISRA Fixes for mlan_glue.c patch 1
[Issue]
Rule 10.1 Operands shall not be of an inappropriate essential type.
Rule 10.3 The value of an expression shall not be assigned to an object with a narrower essential type or of a different essential type category.
Rule 10.4 Both operands of an operator in which the usual arithmetic conversions are performed shall have the same essential type category.

[Fix]
Comply as per rules

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.
STA iperf TCP UDP Client Server

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

Signed-off-by: Abhishek Misra <abhishek.misra@nxp.com>