Issue: WSW-20644 [RT1170][9177][UNII-4] Not able to scan and connect Firecrest-STA to Firecrest-uAP configured in UNII-4 band channels 173 and 177.
[Issue]
STA was not able to connect to AP on Channel 173 and 177.

[Issue Analysis]
AP cannot start on DFS channel 173 and 177 due to which STA not able to
scan and connect.

[Fix]
Added macro (IW61x) for Firecrest for US UNII-4 bands.
Also set the correct region code using the enhanced cli.

[Validation done]
Setup the AP using cli and connect the STA with AP for corresponding
channels.
Able to associate and connect to the desired channels.

[Open Issues/Todo]
None

[Reviewers]
Abhinav Kulkarni, Sandip Dalvi, Supriya Inamdar, Sanjay Yadav.

Signed-off-by: Krishna Singh <krishna.singh@nxp.com>