Issue: WSW-23367 Set 802.11 AX OBSS Narrow Bandwidth RU Tolerance Time.
[Description]
In uplink transmission, AP sends a trigger frame to all the station that will be involved in the upcoming transmission, and then these
stations transmit Trigger-based(TB) PPDU in reponse to the trigger frame.
If STA connects to AP which channel is set to 100,STA doesn't support 26 tones RU.

Pls set this value before connecting to uAP.
The range of Tolerance Time is [1...3600].
Tolerance Time is in unit of second.
STA periodically check AP's beacon for ext cap bit79 (OBSS Narrow bandwidth RU in ofdma tolerance support)
and set 20 tone RU tolerance time if ext cap bit79 is not set.

Signed-off-by: Qiankun Li <qiankun.li@nxp.com>