Issue: WSW-23893 Fix IPv6 Link-local address not change after setting new MAC address in MMH mode
Signed-off-by: xinyuhong <xinyu.hong@nxp.com>
[Description]
IPv6 Link-local address not change after setting new MAC address in MMH mode.
[Analysis]
From the code found that after receiving the event Wifl EVENT UAP STARTED in the uap state machine process, the link local address of the uap mode is not generated based on the new mac address in the net configure address, but in the sta mode the link-local address is generated normally.
[Fix]
generate the new link-local address as the sta mode.
The fix is only verified on RW610