Issue: WSW-24175 Add the hook function based on ipv4 route src
Signed-off-by: xinyuhong <xinyu.hong@nxp.com>
[Description]
RW610 as uap, third-party mobile phone as STA, multicast iperf verification is failed when uap as the iperf client. The reason why the verification fails is that the corresponding netif cannot be found, so the iperf stream cannot be sent. According to the code, in uap mode, searching for netif in ip4_route will use netif_default as the final choice, while netif_default is NULL in uap mode;
[Fix]
Define corresponding interface lwip_hook_ip4_route_src(src, dest) of the hook interface LWIP_HOOK_IP4_ROUTE_SRC(src, dest) in lwip to find the corresponding netif. This interface find netif based on src, src is the IP address carried by the -B parameter of the iperf command.