Issue: WSW-24570 [RT1170] [9177] [STA] "Error in sending Background traffic" messages seen on DUT while running (WMM)Background TX-Traffic from DUT to AP.
Tx port is going out of sync resulting in no port availability for some packets. Thsi is seen with non WMM case as well and this behavior is possible.
Guarding the error messages so they are not seen regularly.

Reviewer: Sandip D, Sanjay Y.
Tests: Tested the traffic for all categories and it is working.

Signed-off-by: Abhinav Kulkarni <abhinav.kulkarni@nxp.com>