Issue: WSW-27775 [Pre-Cert][RT1170][Firecrest_W9177][11n][STAUT]-20/40 BSS coexistence management is not supported in association request.
[Fix]
In earlier implementation, the BSS CoEx management bit was set only if the AP supported the same (in its' beacon frame). As per the discussion with QA team, this behaviour is changed. Now the CoEx bit is set always.

[Validation done]
Verified the CoEx bit is set in the Assoc request for 2.4GHz and 5GHz

[Open Issues/Todo]
None.

Signed-off-by: Sandeep Patil <sandeep.patil@nxp.com>