Issue: WSW-27554 memory leak during scan of wps session.
[Issue]
If trigger sta wps pbc, and don't push PBC button on EXT-AP, a few seconds later, the memory error log is shown on console.

[Fix]
(1) If the scan table is full, free ies of the lowest rssi entry before this entry is replaced
(2) If the scan table is full, free ies of the new entry with lowest rssi, which won't be added into table
(3) Free ies of the old entry if it's duplicate entry
(4) Free ies of the entry, which is error parsing/interpreting the scan response

[Validation done]
After mutiple wps session, and no memory leak.

[Open Issues/Todo]
None

[Reviewers]
Gannicus Guo, Abhinav Kulkarni, Supriya Inamdar.

Signed-off-by: Sandip Dalvi <sandip.dalvi@nxp.com>