Issue: WSW-29354 Redfinch Station fails to get IPv4 from DHCP server during roaming in WPA2 and WPA3
[Issue]
If the signal of AP is very low, after receiving 0x2b event and during the DHCP requesting, driver may receive 0x03(link lost) event from FW.
Then the re-scan is triggered, when the dhcp timer expires, and if the the wlan.sta_ipv6_state is already CM_STA_CONNECTED,
the timer callback will configure the wlan.sta_state to CM_STA_CONNECTED, and the scan results won't be handled.
[Fix]
Stop the dhcp timer first after link lost occurs, as the dhcp timer callback may lead to that the connection state is out-of-sync with FW.
[Validation done]
The issue isn't reproduced.
[Open Issues/Todo]
None
[Reviewers]
Gannicus Guo

Signed-off-by: maochen <maochen.wang@nxp.com>

Signed-off-by: Vikram Hande <vikram.khande@nxp.com>