Issue: WSW-30140 Add Fixes to WM_OS Layer for Zephyr
[Issue]

The WM_OS porting layer for Zephyr implements os_event_notify calls
incorrectly because k_wakeup() has no effect if the target task is not sleeping.
Therefore, if a task notification is set pending via k_wakeup() while the
other task is not blocked, then when the task becomes blocked with
k_sleep(), it will not consume this notification and will be blocked
until another k_wakeup() call with its task ID.

[Fix]

Reimplement the task OSA layer for Zephyr. The os_thread_t type now
defines a structure, which contains a k_tid type as well as a semaphore.
k_thread_custom_data_set() is used to store a pointer to this structure
within every thread, so os_get_current_task_handle() is still
functional. The os_event_notify functions now make use of this semaphore
to emulate FreeRTOS style task notifications.

[Validation done]

This fix was tested with the IW416 module running on an RT1060 EVKB with
the Zephyr supplicant port. I verified that the WiFi module responds to
commands, and was able to scan for networks. Note that the adapter only
initializes correctly from a power on reset.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com
Signed-off-by: Sandip Dalvi <sandip.dalvi@nxp.com>