Issue: WSW-28765 [RT1170] [9177] [STA] [Wpa_supplicant] wlan_get_beacon_period() API returns Ex-AP Beacon period as 0 on DUT when the Ex-AP beacon period is configured to 100.
[Issue]
During connect to AP, the STA first scans for APs in neighbourhood and then initiates connection with the selected AP. At this point, the AP parameters obtained during scanning need to be copied/stored with wpa-supplicant.
Added a CLI 'wlan-get-beacon-interval' to retrieve the beacon-interval of the connected-AP for station.

[Validation done]
Verified that the correct beacon-period is returned via the CLI

[Todo]
None

Signed-off-by: Sandeep Patil <sandeep.patil@nxp.com>