Issue: WSW-25713 [FreeRTOS][Redfinch] Fix UDP RX low TP issue in the multiple test
[Issue]
Run UDP RX for the first time, the TP is 100Mbps. Then run for the second time, the TP will drop to 8Mbps in sheild-room.
In the latest code base, the UDP ptimer will be stopped when the test is done.
The scheduling between imu task and tcp/ip task relied on the 1 ms UDP timer before, and the UDP timer is designed for UDP TX.
Now stop the UDP timer after the first test, then in the second test, the imu task will take the CPU for more time,
and it will fail to malloc memroy from MEMP_TCPIP_MSG_INPKT when sending msg to tcp/ip task.

[Fix]
Call portYIELD() to yield task, when malloc memroy failed from MEMP_TCPIP_MSG_INPKT.

[Validation done]
The TP of multiple times:
  UDP-RX   TCP-RX
  100Mbps  47Mbps

[Open Issues/Todo]
None

[Reviewers]
Gannicus Guo

Signed-off-by: maochen <maochen.wang@nxp.com>