Issue: WSW-32771 Revert "Issue: WSW-32771 fix coex supplicant wakeup fail issue"
[Issue]
Issue: coex_app_supp_cli stress test fail, the wakeup timeout is small.
when the fw is sleeping, the wpa_supplicant task sends auth and send 0x10c command subscribe mgmt frame, and the driver wants to wake up the fw by pmu.
before the driver wakes up the fw, the driver get the sleep_rwlock and set the timeout is 3s.
the fw resp EVENT_PS_SLEEP to driver over 3s, and resulting in the driver 3s timeout and entering into assert.
after timeout, the wpa_supplicant task print the fail log, but the coex print function is nonblocking and sleep,
at the same time, the fw send the EVENT_PS_SLEEP, the cpu switch to the wifi_driver_main_loop task, and release the sleep_rwlock,
so we can see the sleep_rwlock is release but the driver wakeup fw timeout.

[Fix]
modify the wakeup timeout from 3 to 5, but it is better for fw to fix the issue

[Validation done]
coex_app_supp_cli stress test, the timout can't recurrent

[Open Issues/Todo]
None

[Reviewers]
Gannicus Guo, Sandip Dalv

This reverts commit ac63149327e6479c70852efae332ab4721fa6f8e.