Issue: WSW-32297 Add new command wlan-connect-opt to support skip DFS configuration.
[Issue]
If connecting to an WPA2 AP, it will take 10 seconds for the whole
connection process. To reduce the connecting time, added new command so
that user can skip DFS channels when scanning for this connection.

[Fix]
New command is wlan-connect-opt <profile> skipDFS. Once skipDFS is
added, driver will skip DFS channels when assembling channel lists for
scan. This parameter will be clear when connecton process is done,
regardless of connection is successful or not.

[Validation done]
1. Compilation is ok.
2. Verified with both wifi_cli and wifi_wpa_supplicant examples. If
   configured to skip DFS, then channel list in 0x107 commands didn't
   include DFS channels. Once connection is done, the skip DFS setting
   was cleared.

[Reviewers]
Gannicus Guo, Sandip Dalvi, Maochen Wang, Fengming Ye

Signed-off-by: Hui Bai <hui.bai@nxp.com>