Issue: WSW-20166 [RT-1170] [Firecrest] [Tx-Power][UL-OFDMA]: Not able to load RU tx power limit with RT.
[Issue]
Not able to load RU tx power limit with RT.

[FIX]
The wifi_d() debug messages were creating too much debug dump. It is not
needed in normal situations, hence put it behind an #ifdef switch.
This fix does not directly affect the original issue WSW-20166, but it is clubbed with
it for consistency.

[Validation done]
Build done and the image do not show the debug-prints anymore

[Open Issues/Todo]
None

Signed-off-by: Sandeep Patil <sandeep.patil@nxp.com>