Issue: WSW-34605 wifi_cli roaming fail
Issue: the roam_reassoc flag is set to false in the bg scan report, and
the driver can't enter reassoc.
when the rssi low event comes on, the roam_reassoc should set to true,
and in the bg scan report we shouldn't set the roam_reassoc to false. In
the handler_scan_result function, when best ap is the connected ap, the
roam_reassoc will be set to false, and when finding the match ap and
start assocation, the
roam_reassoc should be set to true.

[Fix]
remove the code that set roam_reassoc flag to fasle in the bg scan
report

[Validation done]
wifi_cli validate
wifi_wpa_supplicant validate

[Open Issues/Todo]
None

[Reviewers]
Gannicus Guo, Sandip Dalv