Issue: WSW-35641 [Redfinch][NCP][UART][Supplicant]Fix can config region code successfully after uAP start
[Issue]
[Redfinch][NCP][UART][Supplicant]can config region code successfully after uAP start
[Analyze]
When uap is started, modification of region_code should not be allowed. This restriction is missing in the code of ncp example.
[Fix]
Add a judgment on whether uap has been started in the code that modifies the region code. If so, the region code cannot be modified.
[Validation done]
Validation passed ncp_bridge and ncp_bridge_supp.
[Open Issues/Todo]
None
[reviewers]
Gannicus Guo
Signed-off-by: xinyuhong <xinyu.hong@nxp.com>