Issue: WSW-37320 [RT1060][9177][STA][5G/2.4G][Wpa_Supplicant] 20-25% Throughput drop is observed in HE(80MHz & 40MHz)TCP-RX compared to VHT TP numbers.
[Issue]
Here traffic was going with MCS 0xb which is causing TP drop since RT
couldnt send more number of TCP acks with in timelimit of MCS 0xb
aggregation in firmware.

[Fix]
Fixed this by introducing TID value for TCP ack packets for 80 and 40M BWs. This results in ack going with MCS 9.

[Validation done]
Tested traffics for 20, 40 and 80M BWs using WPA supplicant and TP is as
expected as of VHT.

[Open Issues/Todo]
None

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Sanjay Y.

Signed-off-by: Abhinav Kulkarni <abhinav.kulkarni@nxp.com>