Issue: WSW-38503 [FC-RT1060-EVKC][WPA Supplicant][uAP] Print of Number of Clients connected to APUT is wrong
[Issue]
Print of Number of Clients (STA's) connected to APUT (when "wlan-get-uap-sta-list"
command is given) is wrong. It was due to driver relying on FW for printing the
list. From the FW the EVENT_MICRO_AP_STA_DEAUTH (2c) was coming late when client
get disassociated which was causing the list to be updated late.

[Fix]
Added code in driver to use list of connected client from hostapd directly instead
of getting it from FW.

[Validation done]
Tested wlan-get-uap-sta-list and checked output matches with actual number of
clients connected.

[Open Issues/Todo]
None

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>