Issue: WSW-44713 [RW610] Unsupported STBC, LDPC and BW should not be config in fix rate cmd
[Description]
For this JIRA issue, the invalid BW was input for 0xd6 cmd, so caused array bound issue in FW.
RW610 does not support STBC and LDPC, and it's 20MHz only chip, so it should not set STBC and coding
as 1 in 0xd6 fix rate cmd, and the bandwidth can only be 20MHz.

[Fix]
For RW610, if the invalid STBC, LDPC or BW was input for this cmd, directly return rather than continue
to build and download this 0xd6 cmd.

[Test]
After change, when input invalid STBC, LDPC or BW for RW610, driver can directly return and print corresponding
error log, and the cmd timeout issue of this JIRA was also gone.

[Potential Side-effect]
NA.

[Porting To]
NA.

Signed-off-by: Weijian Shao <weijian.shao@nxp.com>