Issue: WSW-48912 [Redfinch][wifi_cli_static][MMH]Fix dhcp timeout when connecting DUT uAP
[Description]
Fix dhcp timeout when set regioncode with CN.
When the regioncode set to CN, and set uap ssid like frdmap, it will cause issue that sta can't get ipv4 from uap because of dhcp timeout.
[Root Cause]
Through the sniffer log, it is found that aci_aifsn.acm is set to 1 in the beacon of the uap, which causes the sta connected to it to be unable to send data packets normally, so the sta's DHCP discover is dropped by the fw, resulting in DHCP timeout.
aci_aifsn.acm is not assigned a value during the uap configuration process, but the default value of the address corresponding to the cmd buf is used, which may cause the cmd buff to have a dirty value, that is, the value of the address where acm is located may be 1.
[Fix]
During the uap config process, assign the value of aci_aifsn.acm to bss->param.bss_config.wmm_para.ac_params[ac].aci_aifsn.acm to ensure that the correct acm is sent to fw.
[Testing]
Test passed with wifi_cli_static.
[reviewers]
Gannicus Guo
Signed-off-by: xinyuhong <xinyu.hong@nxp.com>

Change-Id: Ie2e8d6935a9e140bc7962cfa7d5e15a8a51a786e