Change of the neighbour bit in the association request
[Description]
Customer reported that wpa_supplicant does not set the neighbour bit in
the assoc request in RM capabilities but wifi_cli example sets this bit in
the assoc request.
The wpa_supplicant (reference)code for this bit is in the function
sme_auth_handle_rrm(), the bit is WLAN_RRM_CAPS_NEIGHBOR_REPORT, which
is not set by default.

[Root Cause]
This bit is set in the wifi_cli code, not set in the wpa_supplicant code

[Fix]
Since wpa_supplicant open source code does not set this bit (not a
mandatory bit), we need to clear this bit in the wifi_cli code as well

[Testing]
Tested locally, RRM related capability works fine

[Potential Side-effect]
NA

Change-Id: 2f4d0d7e-21e7-4d82-9e81-05fa925809d1

Signed-off-by: Chetan Patil chetan.patil@nxp.com