Fix embedded supplicant connect to WPA3 suiteb AP and fail
[Description]
Embedded supplicant does not support WPA3 enterprise security.
But we connect to AP when it claims to support it and fail.

[Root Cause]
wlan_update_rsn_ie adds this type to preference list and
pick it with higher priority than SAE.
So even if user adds network with security type SAE, we still
connects AP with WPA3 suiteb if AP supports it.

[Fix]
Remove all the unsupported security types and remain only WPA2 PSK
and WPA3 SAE for embbedded supplicant.
And add FT security types when CONFIG_11R is enabled.

[Testing]
Compilation successful and sanity test.

Change-Id: I641cb9aa8c0316665b71710a9f37185affa2c9c8
Signed-off-by: Fengming Ye <frank.ye@nxp.com>