Updated lwIP iperf include directive
[Description]
The include of lwiperf.h updated to match
the official include paths of lwIP and to
have the same form as in lwiperf.c in lwIP.
The direct include path to the folder with
lwiperf.h file is going to be removed in lwIP,
so compilation would break without this change.

[Root Cause]
None

[Fix]
None

[Testing]
Compile of wifi_cli ok with both current
version of lwIP and the ongoing one with
the direct include path removed.

Change-Id: 96aca5d2-516e-4f8e-b215-f38b27ed1685

Signed-off-by: Stanislav Poboril <stanislav.poboril@nxp.com>