Keep flexibility to override a choice Kconfig in a Kconfig.defconfig
[Description]
Keep flexibility to override a choice Kconfig in a Kconfig.defconfig

[Root Cause]
A choice can only be overridable in a Kconfig.defconfig if the choice has a "name".
So, if a choice does not have a name then it cannot be overriden in a Kconfig.defconfig.

[Fix]
Add choice name

[Testing]
Test pass

[Open Issues/Todo]
None

[Reviewers]
Gannicus Guo, Maochen Wang, Fengming Ye, Hui Bai

[Port proposal]
Non-general

Change-Id: I98f2e989e216ea60a80af025c36c4b7babfae696
Signed-off-by: Gang Li <gang.li_1@nxp.com>