Fixed Console freeze is seen when second Ex-STA connects to APUT
[Description]
As soon as second STA connects to APUT, Console freeze is seen.
UUT resumes normal operation only after hard reboot

[Root Cause]
PBUF_POOL_BUFSIZE getting short as rx_buf_len is greater then BUFSIZE.
Issues seen with a particular ExSTA SkyHawk.
with this got packets of Rx Len 3328 and this was causing memory corruption, as pbuf size was 2048 then.

[Fix]
This is WAR:
So, to work with this packet size we increased the pbuf size.
with this we are not seeing any of the hang issues, and QA has validated it too.

[Testing]
Validated done with second STA connects to APUT and test is running.

[Open Issues/Todo]
Optimaise the logic and reduce memory consumption.

[Reviewers]
Gannicus Guo, Pankaj Razdan, Sandip Dalvi

Change-Id: b2a2e058-04e4-11f0-a030-00155d89ea3b

Signed-off-by: Manav Chotalia <manavashwinbhai.chotalia@nxp.com>