optimize host sleep interaction with FW
[Description]
We will always notify FW that we woke up after host sleep.
And always notify FW that we are about to sleep.
This will increase average current in periodical host sleep cases,
when wifi is not the wakeup source.

[Root Cause]
We will always notify FW that we woke up after host sleep.
And always notify FW that we are about to sleep.
This will increase average current in periodical host sleep cases,
when wifi is not the wakeup source.

[Fix]
1. Not notify FW when we are not woken up by WLAN.
2. Not do handshake with FW when we didn't notify FW that we woke up.

[Testing]
Compilation successful and sanity test.

Change-Id: Id12a13db6cbe171d6a2cede69c69274757af6305
Signed-off-by: Fengming Ye <frank.ye@nxp.com>
Signed-off-by: Hui Bai <hui.bai@nxp.com>