[IW610 RTOS][MFG] MFG NH is failing with latest RTOS SDK
[Description]
The cpu2 mfg fw not work based on NH 1060EVKC

[Root Cause]
For BLE only app, the cpu2 is download by uart, it should do wifi bt module config before uart init, the process should be disable and then enable, the issue is that enable the module directly, which cause the fw download fail.
During the uart initialization process, for the 1060EVKC+NH chip, the baud rate of the uart between cpu3 and cpu2 should be 3000000. The current configuration value is 115200, which causes the uart to not work properly.

[Fix]
Add the disable module process before enable it.
Change the bandrate from 115200 to 3000000 for 1060EVKC+NH chip

[Testing]
Test with uart_wifi_bridge app.

Change-Id: e0eaabc816ad11f0a81f9f4f2347237c
Signed-off-by: Xinyu Hong <xinyu.hong@nxp.com>