Fix uAP activated fail when STA has connection on DFS channel
[Description]
The SoftAP can't be brought up when STA has connection on DFS channel.

[Root Cause]
When setting up uAP in do_start(), driver will check if the channel is
allowed or not and DFS channels are not allowed for uAP.

[Fix]
Skip the channel check when STA has connection. In this case, uAP must
follow the channel of STA is on.

[Testing]
1. When STA has connection on DFS channel, uAP could be brought up
   successfully on same channel, regardless of configured channel.
2. When no connection on STA, DFS channels are not allowed for uAP.

[Reviewers]
Gannicus Guo, Maochen Wang, Gang Li

Change-Id: I2ba4cbb9f59aa0fac296b431e59679d7f8c2d278
Signed-off-by: Hui Bai <hui.bai@nxp.com>