RSSI and Role info are not correct in wlan-info
[Description]
RSSI and Role information are not correct in wlan-info, when UUT acts as P2P
Client.

[Root Cause]
No logic was present to print info for wfd network. The RSSI command was not
sent to wfd interface.

[Fix]
Added logic for wfd interface in RSSI cmd and role print.

[Testing]
Tested wlan-info after P2P connection.

Change-Id: f39820a4-35a0-11f0-8d75-00155dc95910

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>