Fixed build error for STA only case in wpa_supplicant
[Description]
Build error is seen when softap and hostapd support is disabled.

[Root Cause]
If P2P is enabled and hostapd is disabled than it is causing build error because
P2P is depends on hostapd support.

[Fix]
Disable P2P when Hostapd is disabled in default config.

[Testing]
Build successful and no errors on warning seen.

Change-Id: 72dabe88-36e3-11f0-9ed1-00155dc95910

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>