STAUT-Legacy Roaming | Link lost seen when DUT roaming with UDP traffic
[Description]
During legacy roaming in wpa2 security link lost is seen when DUT is roaming with UDP traffic.

[Root Cause]
This issue is seen because during roaming driver was sending EAPOL (M2 and M4) to fw with same tx priority as UDP traffic i.e., tid as 0.
Due to this DUT was trying to send multiple EAPOL (M2 and M4) frames to fw which was further not able to transmit EAPOL (M2 and M4) frames as TX_MAC_QUEUE of fw is most of the time full of udp traffic that is running in background. As both EAPOL and UDP packet are trying to use same TX_MAC_QUEUE because of tx priority, most of the time would leads to 4WAY_HANDSHAKE_TIMEOUT.

[Fix]
Updated tid value from 0 to 7 for EAPOL packet so as to send EAPOL frame with highest priority.

[Testing]
Executed roaming for around 5 hrs 4WAY_HANDSHAKE_TIMEOUT was not seen.

[Open Issues/Todo]
None.

[Reviewers]
Gannicus Guo, Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Keval Thakker, Sanjay Yadav, Pankaj Razdan.

Change-Id: 601984b6-6166-11f0-93ca-47ac6699422a

Signed-off-by: Mayur Mansharamani <mayur.mansharamani@nxp.com>