Fix legacy roaming fail if STA didn't receive AP's SAE auth commit
[Description]
In legacy roaming (WPA3 SAE security mode), when STA roaming from AP1 to AP2,
STA send out SAE Auth commit and acked by AP2. But AP2 didn't respond Auth commit
to STA, which result in roaming to AP2 failed.
STA scan failed and don't roaming or connect again.

[Root Cause]
In legacy roaming mode, when a bgscan result is received,
it updates the network info (roam to AP2) and associates.
If the association fails, it re-scans. During the re-scan,
an rssi low event is received and wlan.roam_reassoc is set to true.
Later, when a re-scan result is received, it finds a matching AP.
Since wlan.roam_reassoc is true, it checks if the bssid of
the current network is the same as the bssid of best_ap.
They are both the bssid of AP2, it stops associating.

[Fix]
Check if mlan_adap->priv[0]->curr_bss_params.bss_descriptor.mac_address is
same as best_ap's bssid.

[Testing]
Legacy roaming pass

[Reviewers]
Gannicus Guo, Hui Bai

[Port proposal]
general

Change-Id: I06c87f48000b4369c9583e6713d698729c6fd761
Signed-off-by: Gang Li <gang.li_1@nxp.com>