Tx-frame command is failing for HE20MHz MCS10,11
[Description]
Issue was seen due to missong variable in structure. adv_coding was
missing so added it.

[Root Cause]
Issue was seen due to missong variable in structure. adv_coding was
missing so added it.

[Fix]
Issue was seen due to missong variable in structure. adv_coding was
missing so added it.

[Testing]
Tested with RFTM app in failing and working condition

[Open Issues/Todo]
NA

[Reviewers]
Gannicus Guo, Sandip D, Sanjay Y

[Port proposal]
general

Change-Id: 403c201b-ad21-4bb0-af04-3f534ac4ba4b

Signed-off-by: Abhinav Kulkarni <abhinav.kulkarni@nxp.com>