[P2P] P2P client connection fails when connection is made after P2P GO is stopped
[Description]
Start P2P in auto GO mode and then stop the auto GO. After doing this, try to
connect in client mode to another P2P auto GO device. The connection then fails.

[Root Cause]
The issue occurred because the BSS role was not reset in the stop_ap callback
API, which caused an issue in wifi_nxp_send_mlme() during the sending of an
action frame.
In wifi_nxp_send_mlme(), the driver checks if the role is STA for a WFD
interface and then sends the remain-on-channel command to firmware. If this
command is not sent before the action frame, the firmware will drop the packet
and return tx_status_report as failure.

[Fix]
1. Reset the BSS role after uAP is stopped.
2. Send the SYS_RESET command otherwise, firware cannot find the P2P device
   after persistent group removal.
4. Reconfigure the WFD MAC address after the BSS is stopped.

[Testing]
1. Verified that a P2P client successfully connects after auto GO is stopped.
2. Verified that P2P connections work correctly with both P2P and non-P2P
   devices.

Change-Id: c4ceb0d8-72d3-11f0-8af5-00155d847253

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>