[P2P] Enabled WMM if P2P is enabled
[Description]
WMM is a prerequisite for P2P as per the P2P certification specification section
3.1.

[Root Cause]
WMM is not enabled for all SoC by default so if users enabled P2P than it will
lead to connection failure in case of P2P persistent GO.

[Fix]
Enabled WMM when P2P is enabled.

[Port Proposal]
Non-general.

[Testing]
P2P persistent and invitation cases are passing.

Change-Id: a9bf8062-7d9e-11f0-9e50-00155d847253

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>