Align IW610 default wakeup condition setting with RW610
[Description]
The default value of wlan.wakeup_conditions on IW610 is not align with
RW610.

[Root Cause]
The default value of wlan.wakeup_conditions on RW610 is 0 but the value
of IW610 is 0x1e. This will cause unexpected WLAN wakeup if user set MEF
conditions.

[Fix]
Align with RW610, set wlan.wakeup_conditions parameter to 0 by default
on IW610.

[Testing]
Verified ok.

[Reviwers]
Gannicus Guo, Rex Chen

Change-Id: Ib5e2b9d2c6e36cfe79fa920ae6d68d1682ad841f
Signed-off-by: Hui Bai <hui.bai@nxp.com>