fix the out of bunds coverity
[Description]
High impack coverity in function wifi_nxp_wpa_supp_probe_req_report.

[Root Cause]
The number of elements in the 'priv' array is 2, so the indices should be 0 and 1.
Currently, 'bsstype' is set to 2, which causes an array out-of-bounds error.

[Fix]
Increase macro protection for bss_type.

[Testing]
Done

[Reviewers]
Gannicus Guo, Qingling Wu

[Port proposal]
general

Signed-off-by: Pan Gao pan.gao@nxp.com
Change-Id: I8b4fa0aef41e3186c9370b0f8a13159cca5fd624